* Added CalendarIdSetter.jsp to add a Calendar ID to Volunteer objects who don't have an ID stored.
* Modified MakeUserServelet to check for prior existence of the new name
before creating a new user; if name exists, use that User.

* Modified MakeUserServelet to create a new Calendar (setting access to Public)
for the new user, storing the Calendar ID inside the new Volunteer object.

* Modified Volunteer.java to store the ID of the Volunteer's calendar.

* Modified VolunteerCoordinatorServlet.java to call doPost when the doGet
method is called.

* Modified addVolunteerServlet to create the relevant EventEntry and add itto the Volunteer's calendar.

* Modified calendar.jsp to no longer create or edit calendars for the user.
It now only fetches the CalendarID associated with the user from the datastore.
In the event there is no ID for that user in the datastore, it fetches one by
finding a calendar in that user's name and getting the ID.  This should only be
used until we delete all the old stored calendars from the previous system.